exercise (bstree): Insert works! (well it compiles)
Thanks to this gist I was able to get my BSTree `insert` into shape:

https://gist.github.com/aidanhs/5ac9088ca0f6bdd4a370

The trick was to fix my pattern matching and to use `ref` in my pattern
match, to avoid moving into my value. Still processing what that means
exactly, but I'm getting there!

My implementation is a bit verbose and has some copy-paste for the left
and right branch code-paths -- but we can fix that later.